Change datatypes for compatibility with dbweb
Blobs of XML need to be text, and floats need to specify precision of
32 bits for good lat/lng values.

+++ db/migrate/20110603185622_make_node_primary_key_network_dependent.rb
@@ -16,7 +16,7 @@ class MakeNodePrimaryKeyNetworkDependent < ActiveRecord::Migration
         `network_id` integer NOT NULL REFERENCES `networks`,
         `id` INTEGER NOT NULL REFERENCES `node_families`, 
         `name` varchar(255), `description` varchar(255), 
-        `type_node` varchar(255), `lat` decimal, `lng` decimal, 
+        `type_node` varchar(255), `lat` float(32), `lng` float(32),
         `elevation` decimal, `created_at` datetime, `updated_at` datetime, 
         PRIMARY KEY(`network_id`,`id`)
       )
